+++ src/api/login/LoginApi.tsx
@@ -1,19 +1,6 @@
 import axios from 'axios';
 import API_BASE_URL from '../Apiconfig';
 
-const ACCESS_TOKEN = 'Access_token';
-
-const accessToken = localStorage.getItem(ACCESS_TOKEN);
-
-const headers = new Headers({
-  'Content-Type': 'application/json',
-});
-
-if (accessToken && accessToken !== '') {
-  headers.append('Authorization', accessToken);
-  axios.defaults.headers.common.Authorization = `${accessToken}`;
-}
-
 // 인터페이스 정의
 interface UserDto {
   id: string;
@@ -29,6 +16,7 @@ const login = async (userDto: UserDto): Promise<boolean> => {
       const jwt = res.headers.authorization;
 
       if (jwt) {
+        const ACCESS_TOKEN = 'ACCESS_TOKEN';
         localStorage.setItem(ACCESS_TOKEN, jwt);
         result = true;
       }
