+++ src/routes/MainPage.tsx
@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-
+import React, { useEffect, useState } from 'react';
+import { useAtom } from 'jotai';
 import HomePage from '@pages/homePage/HomePage';
 import MyPage from '@pages/mypagePage/MyPage';
 import GroupPage from '@pages/groupPage/GroupPage';
@@ -9,6 +9,8 @@ import MainDeclarationBoard from '@components/main/MainDeclarationBoard';
 import styles from '@styles/main/MainPage.module.css';
 import Header from '../components/common/Header';
 import TabBar from '../components/common/TabBar';
+import findUserInfo from '../api/user/UserFindInfo';
+import userInfoAtom from '../store/userInfo/UserFindInfo';
 
 const tabs = [
   { id: 1, title: '회원정보', component: <HomePage /> },
@@ -19,10 +21,24 @@ const tabs = [
 
 export default function MainPage() {
   const [activeTab, setActiveTab] = useState(1);
+  const [, setUserInfo] = useAtom(userInfoAtom);
 
   const handleTabClick = (id: number) => {
     setActiveTab(id);
   };
+  useEffect(() => {
+    const fetchUserInfo = async () => {
+      try {
+        const userData = await findUserInfo(); // findUserInfo 함수 호출
+        if (userData) {
+          setUserInfo(userData); // userData를 atom에 설정
+        }
+      } catch (error) {
+        console.error('Error fetching user info:', error);
+      }
+    };
+    fetchUserInfo();
+  }, []);
 
   const ActiveComponent = tabs.find((tab) => tab.id === activeTab)?.component;
